Move `TaskStoreSpy` to the test file where it's used
